Change visited pixel criteria to fix a bug
Prior to this change, the way the solving algoritm checked for a
visited pixel was wrong, since it only compared the color of the
current pixel with the visited color.

The previous behavior caused a bug when pressing the "Solve" button
multiple times in succession, since each press interrupts the previous
routine and starts a new one.
As each routine executed, the solution to the maze would be found
and drawn over the image every time.
This is because despite the fact that the algorithm marked pixels as
visited before finding the solution, the solution itself has a different
color than those visited pixels.
Thus the solution was found once for each button press, instead of
only once.
Due to this the algoritm would unnecessarily:
    - Identify the solution path as unvisited
    - Mark it as visited, overriding its color
    - Find the solution again
    - Mark the solution with its color

With this change, the algorithm now compares against both the visited
color and the solution color, fixing the previously mentioned bug.