Refactor application
The relevant variables of the application have been moved into the
ApplicationState class, and they are now shared across the controllers.

All the validation of the values in the ApplicationState has been moved
into its own module validation.py

Thanks to this, now every individual controller can be instantiated
with only an ApplicationState instance and a Validator instance.